Factored out AtB personnel market method.
Moved to a separate package that will eventually include all AtB code.
